file name and additional deletion check
